Continue the hack job of refactoring terms.

+++ Clasp/Data/Terms/FixNum.cs
@@ -4,9 +4,9 @@ namespace Clasp.Data.Terms
 {
     internal readonly struct FixNum : ITerm, IEquatable<FixNum>
     {
-        public readonly long Value;
+        public readonly int Value;
 
-        public FixNum(long value) => Value = value;
+        public FixNum(int value) => Value = value;
         public bool Equals(FixNum other) => Value == other.Value;
         public bool Equals(ITerm? other) => other is FixNum fix && Equals(fix);
         public override bool Equals(object? other) => other is FixNum fix && Equals(fix);
